OpenCV: 4주차(2) 실습, 3주차 과제

+++ OpenCV/ch04/01/04.event_mouse_trackbar.py
@@ -5,7 +5,7 @@ import cv2
 def onChange(value):
     global image, title  # 전역 변수 참조
 
-    add_value = value - int(image[0][0])
+    add_value = np.uint8(value - int(image[0][0]))
     image = image + add_value
     cv2.imshow(title, image)
